Added slider to control speed (see desc)
The simulation prefers accuracy over speed, eg if your computer cannot
handle a certain speed, it will not go over it to prevent simulation
errors.

Also, there is a bug where if you set the speed really high and then
back down, depending on how long it was up high, it will take a while to
actually slow down, due to the way dyn4j works.