Fixed bugs (see description)
I fixed the bug where if you sped up the simulation for a while and then
back down, it would take a while to slow back down. I fixed it by using
world.setAccumulatedTime(0) to reset the world's timer so it would stop
doing extra steps when it shouldn't be.

I also fixed the bug where the top 50% creatures would not duplicate
correctly.

Each creature keeps track of the original position of all of its nodes.
When I was mutating/duplicating creatures, I was doing something like
this:

Vector2 trans = something;
nodes[i].getTransform.setTranslation(trans);
oPos[i] = trans;

Therefore the oPos was the same thing as the translation of the node, so
it would be moved slightly until the game set the translation again in
the update loop.

That bug would cause the top creatures to do worse than they did last
time, even though they should theoretically be identical.

Anyway, yeah.